+++ ldr_tools_blender/__init__.py
@@ -5,7 +5,7 @@ bl_info = {
     "name": "ldr_tools_blender",
     "description": "Import LDraw models in .mpd .ldr, and .dat formats",
     "author": "ScanMountGoat (SMG)",
-    "version": (0, 4, 1),
+    "version": (0, 4, 2),
     "blender": (4, 1, 0),
     "location": "File > Import",
     "warning": "",